refactor: add spending package to add package
