+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
         </activity>
 
         <activity
-            android:name=".presentation.add.AddActivity"
+            android:name=".presentation.add.spending.AddActivity"
             android:windowSoftInputMode="adjustResize">
 
         </activity>
