+++ app/src/main/java/kr/ksw/mybudget/presentation/add/spending/AddActivity.kt
@@ -1,4 +1,4 @@
-package kr.ksw.mybudget.presentation.add
+package kr.ksw.mybudget.presentation.add.spending
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -7,8 +7,8 @@ import androidx.activity.viewModels
 import androidx.compose.material3.Surface
 import dagger.hilt.android.AndroidEntryPoint
 import kr.ksw.mybudget.domain.model.spending.SpendingItem
-import kr.ksw.mybudget.presentation.add.screen.AddSpendingScreen
-import kr.ksw.mybudget.presentation.add.viewmodel.AddSpendingViewModel
+import kr.ksw.mybudget.presentation.add.spending.screen.AddSpendingScreen
+import kr.ksw.mybudget.presentation.add.spending.viewmodel.AddSpendingViewModel
 import kr.ksw.mybudget.presentation.core.keys.SPENDING_ITEM_KEY
 import kr.ksw.mybudget.ui.theme.MyBudgetTheme
