+++ app/src/main/java/kr/ksw/mybudget/presentation/add/spending/viewmodel/AddSpendingUIEffect.kt
@@ -1,4 +1,4 @@
-package kr.ksw.mybudget.presentation.add.viewmodel
+package kr.ksw.mybudget.presentation.add.spending.viewmodel
 
 sealed class AddSpendingUIEffect {
     data object FinishAddScreen : AddSpendingUIEffect()
